Change the name of the types

+++ src/test/java/ru/netology/stats/StatsServiceTest.java
@@ -8,8 +8,8 @@ public class StatsServiceTest {
     @Test
     public void allSalesSum() {
         StatsService service = new StatsService();
-        int[] sale = new int[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int actual = (int) service.allSalesSum(sale);
+        long[] sale = new long[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long actual = service.allSalesSum(sale);
         long expected = 180;
         assertEquals(expected, actual);
     }
@@ -18,17 +18,17 @@ public class StatsServiceTest {
     public void saleAverageMonth() {
         StatsService service = new StatsService();
 
-        int[] sale = new int[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        double actual = (double) service.saleAverageMonth(sale);
-        double expected = 15;
+        long[] sale = new long[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long actual = service.saleAverageMonth(sale);
+        long expected = 15;
         assertEquals(expected, actual);
     }
 
     @Test
     public void maxSale() {
         StatsService service = new StatsService();
-        int[] sale = new int[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int actual = (int) service.maxSale(sale);
+        long[] sale = new long[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        int actual = service.maxSale(sale);
         int expected = 8;
         assertEquals(expected, actual);
     }
@@ -36,8 +36,8 @@ public class StatsServiceTest {
     @Test
     public void minSales() {
         StatsService service = new StatsService();
-        int[] sale = new int[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int actual = (int) service.minSales(sale);
+        long[] sale = new long[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        int actual = service.minSales(sale);
         int expected = 9;
         assertEquals(expected, actual);
     }
@@ -45,8 +45,8 @@ public class StatsServiceTest {
     @Test
     public void saleBelowAverage() {
         StatsService service = new StatsService();
-        int[] sale = new int[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int actual = (int) service.saleBelowAverage(sale);
+        long[] sale = new long[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        int actual = service.saleBelowAverage(sale);
         int expected = 5;
         assertEquals(expected, actual);
     }
@@ -54,8 +54,8 @@ public class StatsServiceTest {
     @Test
     public void saleAboveAverage() {
         StatsService service = new StatsService();
-        int[] sale = new int[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int actual = (int) service.saleAboveAverage(sale);
+        long[] sale = new long[]{8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        int actual = service.saleAboveAverage(sale);
         int expected = 5;
         assertEquals(expected, actual);
     }
